+++ js/filter.js
@@ -1,28 +1,22 @@
+import {reRenderMarkers} from './map.js';
+
 
 const mapForm = document.querySelector('.map__filters');
 const houseType = mapForm.querySelector('#housing-type');
 
-// typeHouses - приходит с сервера
-// housingType - варианты по фильтру
-
 function addFilterListener(offers) {
-
-
   houseType.addEventListener('change', function () {
-    let newOffers = []
 
-    offers.forEach((item) => {
-      const typeHouses = item.offer.type;
-
-      if (typeHouses == houseType.value) {
-        newOffers = houseType.value;
-
-      }
-    });
-
-    console.log(newOffers)
-
-  })
+    // создаем коллекцию подходящих элементов
+    if (houseType.value === 'any') {
+      reRenderMarkers(offers);
+    } else {
+      const filteredOffers = offers.filter((item) => item.offer.type === houseType.value);
+      reRenderMarkers(filteredOffers);
+    }
+  });
 }
 
-export {addFilterListener}
+export {
+  addFilterListener
+}
